ts_config was accidentally merged to the wrong directory
